Move FakeRedis and FakeQueue to conftest.py; adapt test_queue_manager_timeout.py
